+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 🛠 Statistics & Combat Basics and Promotion : Monad Eva guides now available (2025-06-09)
+- Statistics & Combat Basics : Fundamental systems, mechanics, and beginner-friendly guides that apply to the entire game. Learn More [here](/guides/general-guides/stats)
+- Promotion Monad Eva : Boss strategy guide. Learn More [here](/guides/adventure-license/prom-meva)
+
+
 ## 🛠 Walking Fortress Venion update (2025-06-04)
 - Guide is up to date for june 2025 version. Learn More [here](/guides/world-boss/venion)